+++ src/module/data-models/item/abstract/base.ts
@@ -6,6 +6,15 @@ export default abstract class SmtBaseItemData extends foundry.abstract
 
   abstract readonly equippable: boolean;
 
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  static override migrateData(source: Record<string, any>) {
+    if ("target" in source && source.target === "combatants") {
+      source.target = "allCombatants";
+    }
+
+    return source;
+  }
+
   static override defineSchema() {
     const fields = foundry.data.fields;